move verifyDocumentType utility to parser
