+++ src/react/parser/index.ts
@@ -113,3 +113,15 @@ export function parser(document: DocumentNode): IDocumentDefinition {
   cache.set(document, payload);
   return payload;
 }
+
+export function verifyDocumentType(document: DocumentNode, type: DocumentType) {
+  const operation = parser(document);
+  const requiredOperationName = operationName(type);
+  const usedOperationName = operationName(operation.type);
+  invariant(
+    operation.type === type,
+    `Running a ${requiredOperationName} requires a graphql ` +
+      `${requiredOperationName}, but a ${usedOperationName} was used instead.`
+  );
+}
+
